refactor: rename server module to registry
