+++ nebula_common/src/registry/endpoints.rs
